+++ src/VideoStream.cpp
@@ -514,7 +514,7 @@ bool VideoStream::setupFiltering()
     int64_t conversionAlgorithm = SWS_POINT; // optimal speed scaling for videos
 
     char sws_flags_str[128];
-    snprintf(sws_flags_str, sizeof(sws_flags_str), "flags=%" PRId64, conversionAlgorithm);
+    snprintf(sws_flags_str, sizeof(sws_flags_str), "flags=%d", (int) conversionAlgorithm);
     graph->scale_sws_opts = av_strdup(sws_flags_str);
 
     // INPUT BUFFER
